refactor - move logic to different file
